feat(api): BILL-1 catalog — fee heads, bill_fee_structures, discount reasons, transport routes, tax rates, late fee rules
Phase 2 of the billing ledger rebuild (BILL-SPEC.md §5). Adds the new
catalog layer alongside the old finance tables, which keep running
completely untouched.

- Migration 0019: fee_heads, bill_fee_structures/bill_fee_structure_items,
  discount_reasons, transport_routes, tax_rates (ships empty per R6),
  late_fee_rules. bill_fee_structures uses UNIQUE NULLS NOT DISTINCT
  (Postgres 15+) rather than a plain UNIQUE, since a bare unique constraint
  would let two identically-named whole-class structures both slip through
  (every NULL section_id is distinct from every other NULL in ordinary
  SQL uniqueness). Applied to all 8 tenants + tenant_bill_scratch.
- public.tenants gains tagline/paymentInstructions/qrImageUrl (R12).
- Six new services + BillCatalogController: /finance/fee-heads,
  /finance/discount-reasons, /finance/transport-routes, /finance/tax-rates,
  /finance/late-fee-rules, /finance/bill/fee-structures. TaxRateService
  rejects overlapping effective-date ranges (service-layer check, matches
  the codebase's existing pre-check-before-insert convention).
- BUGS-3: bill_fee_structures/bill_fee_structure_items and the
  /finance/bill/fee-structures route are both deliberately not the bare
  spec names — the old table and route are still live. Renamed to the
  canonical names only at the Phase 4 cutover, after the old ones drop.

Two significant findings from live-proofing, both in BILL-BUGS.md:

- INC-3: Phase 1's @IsMoneyString() DTO migration left several raw SQL
  sites without the ::numeric cast Postgres requires for a text-bound
  parameter against a NUMERIC column (a JS number needs no such cast,
  which is why Phase 1's mocked tests never caught it). This had already
  broken the OLD, live PATCH /finance/fee-structures/:id/items and
  setStudentFeeAssignment endpoints, not just new Phase 2 code. Swept the
  entire finance module explicitly for every @IsMoneyString() consumer and
  every remaining raw INSERT/UPDATE; fixed and live-re-proved all nine
  individual write sites across the six affected files against real
  Postgres, using fresh throwaway fixtures (never mutating shared/existing
  rows without capturing their prior state first).
- INC-2 / FIX-QUERY-DTO: the global ValidationPipe's `whitelist: true`
  silently strips undecorated @Query() DTO properties, so the four
  pre-existing finance query DTOs (fee-category, fee-structure, invoice,
  payment) have never actually filtered or paginated over real HTTP —
  invisible until now because every existing test bypasses the pipe by
  calling the service directly. Fixed only the six new Phase 2 query DTOs;
  logged the pre-existing ones as their own backlog item for separate
  prioritization, left untouched.

768/768 tests passing (93 suites), tsc --noEmit clean.

+++ apps/api/src/modules/finance/fee-structure.service.ts
@@ -49,7 +49,7 @@ export class FeeStructureService {
           `INSERT INTO fee_structure_items
              (fee_structure_id, fee_category_id, amount, due_day_of_month, due_date,
               fine_per_day, grace_period_days)
-           VALUES ($1::uuid, $2::uuid, $3, $4, $5::date, $6, $7)`,
+           VALUES ($1::uuid, $2::uuid, $3::numeric, $4, $5::date, $6::numeric, $7)`,
           structure.id,
           item.feeCategoryId,
           item.amount,
@@ -169,7 +169,7 @@ export class FeeStructureService {
           `INSERT INTO fee_structure_items
              (fee_structure_id, fee_category_id, amount, due_day_of_month, due_date,
               fine_per_day, grace_period_days)
-           VALUES ($1::uuid, $2::uuid, $3, $4, $5::date, $6, $7)`,
+           VALUES ($1::uuid, $2::uuid, $3::numeric, $4, $5::date, $6::numeric, $7)`,
           id,
           item.feeCategoryId,
           item.amount,
